Revert mac address of interface when container is stopped.
The interface used inside the container will have its mac
changed often; each time its brought back up. When the
container is stopped it would be nice if the mac address
of the used interface is returned back to its permanent
address. This will help with any host udev rules that
might already be in place.